requirements.txt
psycopg2-binary
requests
python-dotenv
uvicorn
mcp[server]>=1.0.0
pydantic
pydantic-settings
openai>=1.3.0
httpx
starlette
cryptography>=41.0.0
argon2-cffi>=23.1.0
